HACK: bl31_setup: plat_k3_mmap: Distinguish AM62L vs K3 mem maps
Fixme: Identify the required regions for AM62L that need to be mapped and
add them one by one selectively.

HACK to use the mailbox config to distinguish between K3 and AM62L and
build accordingly. We need to find a clean way in upstream to handle
this difference in memory maps

Change-Id: If19214e2f2e31510914d40916c63416fb33e72c7
Signed-off-by: Dhruva Gole <d-gole@ti.com>