HACK: feat(ti): lpm: Add macro to disable AM62L LPM
When LPM is enabled in AM62L the bl31 size increases by 100Kb. Which
this it goes beyound the ROM's capability to load a component in emmc
boot. Add a mcaro to disable the AM62L LPM so that the bl31 will fit in
the ROM's limit for single component.

Change-Id: Ic521c16f3ae1f30366a05af5aa014dd6b72f326b
Signed-off-by: Sebin Francis <sebin.francis@ti.com>